// src/components/editor/models/ECodeTheme.ts
export enum ECodeTheme {
  AMY = 'amy',
  AYU_LIGHT = 'auy-light',
  BARF = 'barf',
  BESPIN = 'bespin',
  BIRDS_OF_PARADISE = 'birds-of-paradise',
  BOYS_AND_GIRLS = 'boys-and-girls',
  CLOUDS = 'clouds',
  COBALT = 'cobalt',
  COOL_GLOW = 'cool glow',
  DRACULA = 'dracula',
  ESPRESSO = 'espresso',
  NOCTIS_LILAC = 'noctis-lilac',
  ROSE_PINE_DAWN = 'rose-pine-dawn',
  SMOOTHY = 'smoothy',
  SOLARIZED_LIGHT = 'solarized-light',
  TOMORROW = 'tomorrow'
}